Warnings-- : I _think_ all the KDE platforms have long at least the size of a pointer. int doesn't cut it, it's only 32 bits on most of our 64-bit platforms. Perhaps we need to add a operator << for pointers like printf's %p.
svn path=/trunk/kdepim/; revision=346949

+++ actionmanager.cpp
@@ -1207,7 +1207,7 @@ void ActionManager::toggleResourceButtons()
 {
   bool visible = mResourceButtonsAction->isChecked();
 
-  kdDebug(5850) << "RESOURCE VIEW " << int( mResourceView ) << endl;
+  kdDebug(5850) << "RESOURCE VIEW " << long( mResourceView ) << endl;
 
   if ( mResourceView ) mResourceView->showButtons( visible );
 }
